chromeos-base: Enable loose reverse path filtering
When policy rules are used to route traffic over a builtin VPN (OpenVPN
or L2TP/IPsec), the kernel's FIB logic will use tun0 as the default
egress interface for all traffic that isn't destined for the VPN gateway.
If strict reverse path filtering is enabled, this means that inbound
ARP requests coming from the LAN router will be rejected
(since tun0 != wlan0), and the network will become unreachable when the
ARP entry for the Chromebook expires.

Since the current plan is to continue expanding the use of per-device
routing tables in order to support multi-routing APIs in Android, we will
need to keep loose filtering enabled indefinitely until we come up with
a way to make it work in this configuration.

BUG=chromium:797058
TEST=manually verify using OpenVPN and L2TP/IPsec

Change-Id: Ib19f198f10fac792211354043542b02c5648ed9f
Reviewed-on: https://chromium-review.googlesource.com/851085
Commit-Ready: Kevin Cernekee <cernekee@chromium.org>
Tested-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>