gdb: add a hack to smooth ncurses upgrade
The current ncurses version is using SLOT=0/5.  That means the ABI is
at "5".  The next version is using SLOT=0/6 since the ABI (SONAME) has
changed.  For installed packages, portage can handle this gracefully
due to the subslot operator.  Gdb is a little tricky though because we
only build it from binpkgs atm, and when it's built against 0/5, the
RDEPEND in the binpkg ends up recording "0/5" instead of "*/5" (as we
have no way of expressing the latter currently).

Add an explicit RDEPEND to gdb so we can work with either version.
This lets existing systems upgrade smoothly in place.  Once we upgrade
everything to version 6, we can revert this back to the way it was.

BUG=chromium:800049
TEST=emerging from source produces a RDEPEND that works w/old ncurses slot

Change-Id: I0901c02b28027ee4c89aad250feee98c24a6878b
Reviewed-on: https://chromium-review.googlesource.com/854821
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>