infineon-firmware-updater: Trigger via flag file.
Remove the mechanism that stored user consent in VPD and instead
consider invocation of the script sufficient evidence that we actually
want to install a TPM firmware update if possible. This allows a quick
check in the init script, thus reducing perf impact to the absolute
minimum for the case where there is no update.

BUG=chromium:780817, chromium:768342
CQ-DEPEND=CL:860379,CL:824609
TEST=unit tests

Change-Id: Ib378c9b2740ae8b7d469b7952003cd89e280d89b
Reviewed-on: https://chromium-review.googlesource.com/857496
Commit-Ready: Mattias Nissler <mnissler@chromium.org>
Tested-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>