perf: v4.14: CrOSify and adapt to the actual version
The current v4.14 ebuild is a copy of gentoo's ebuild for perf v4.9.13,
adapt it to CrOS and v4.14.

As of now perf v4.14 is always built with v4.4 kernel headers. If
features are needed that are only available in newer kernels the
dependency can be adapted after creating the corresponding
linux-headers ebuild.

BUG=chromium:797835
TEST=emerge-kevin/pyro/veyron_minnie perf
  verify v4.14 is built
  cros deploy ${DUT} perf
  run 'perf test'
    => no regressions wrt v4.4

Change-Id: Ib551c9d924d21e519460c80ffbd37ba84ba0d616
Reviewed-on: https://chromium-review.googlesource.com/860908
Commit-Ready: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>