chromeos-cr50-scripts: Disable cr50 firmware update in factory mode.
Update script of cr50 would not perform update when boot in the factory
mode. If firmware should be updated then an explicit factory test should be
arranged to do so. The reason is that a reboot would be occured after
first round of update process and operators in the factory would not
know whether this is a normal behavior or not.

BUG=chromium:800680
TEST=manually test on the DUT

Change-Id: Ibab564b59ca6a093811819c7293ecdf5ad662eb8
Signed-off-by: Marco Chen <marcochen@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/861559
Reviewed-by: Yilun Lin <yllin@chromium.org>