coreboot: Fix symbolized BL31 binary for the serial firmware build
CL:430328 messed up the source path of the bl31.serial.elf output file,
so that it is now identical to bl31.elf when it instead should contain
the code of the serial-enabled build instead. This means that it doesn't
match the code in the final firmware image it was meant to match and
developers that try to use it for debugging will be very confused.

BRANCH=none
BUG=none
TEST=I told Derek to do this and suddenly his bl31.elf wasn't magically
different if you extract it from the final image and compare it with the
unstripped one anymore.

Change-Id: I974a64fcb8d49fe909ca266dd753df4ed7eee3d2
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/879744
Reviewed-by: Simon Glass <sjg@chromium.org>