net-fs/samba: Fix ftruncate to support SMB2
cli_ftruncate was failing for SMB2+ connections. This change fixes
ftruncate so that SMB2+ connections will be able to properly call it.

This bug was previously reported in a Samba channel and a fix has
already been uploaded. The bug was reported on Samba 4.4.8 and the fix
was made live on Samba 4.5.4

Previous bug report: https://bugzilla.samba.org/show_bug.cgi?id=12479
Patch:
https://github.com/samba-team/samba/commit/e0f1ed9f450851bf5b7fec84577b50047309db3f

BUG=chromium:805748
TEST=emerges and ftruncate no longer returns error

Change-Id: I6a1a98010e8fceeabb65170947bf74428c96e6a9
Reviewed-on: https://chromium-review.googlesource.com/884994
Commit-Ready: Allen Vicencio <allenvic@chromium.org>
Tested-by: Allen Vicencio <allenvic@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>