Add an ebuild for chromeos-base/cros-camera-tool
cros-camera-tool is going to be a multifunctional tool for testing
cameras on Chrome OS devices. For now, it is only capable of querying
camera modules in the system, as long as the modules are compliant with
the V4L2 APIs used by MIPI cameras on Chrome OS devices.

Example invocation:
localhost ~ # cros-camera-tool modules list
            Name | Vendor ID
         module1 | 123
         module2 | 456

BUG=b:70357756
TEST=emerge-poppy cros-camera-tool
CQ-DEPEND=CL:890799

Change-Id: Ide0d06929b282725399a8d6130f81653d35c5acc
Reviewed-on: https://chromium-review.googlesource.com/891060
Commit-Ready: Tomasz Figa <tfiga@chromium.org>
Tested-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>