media-libs/mesa: Fix llvmpipe on arm
When our graphics card was llvmpipe and we tried to start the UI, we'd
see in the logs:

LLVM ERROR: Program used external function '__aeabi_unwind_cpp_pr1'
which could not be resolved!

This seems related to clang upgrades and/or the migration to libc++

It was suggested that adding the "-rtlib=libgcc" flag would fix the
problem and, indeed, it did.

BUG=chromium:807103
TEST=See bug

Change-Id: I72748e53f8fe9b6757309fb264c8fd25354cfa8c
Reviewed-on: https://chromium-review.googlesource.com/893540
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>