chromeos-bmpblk: Add vbgfx_not_scaled flag file to CBFS
This patch makes the chromeos-bmpblk.ebuild insert an empty flag file
into CBFS to record when the firmware images were scaled with the
default resolution (1366x768). This file can be detected by the new
firmware_Bmpblk FAFT test, which will alert the operator that this
firmware image isn't ready for release yet.

BUG=None
TEST=Ran firmware_Bmpblk on Kevin with and without triggering this.

Change-Id: Id76628cd82763b3f11232d9d84d8a59f67b12e14
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/898219
Reviewed-by: Hung-Te Lin <hungte@chromium.org>