[GCC] Fix recent break of arm-non-eabi-gcc build.
Recent GCC commit contains a mis-placed #endif in libgcc/emutls.c,
which is causing the arm-none-eabi-gcc build (and therefore the
chromiumos-sdk builder) to fail.  This CL fixes that.  This is a
temporary fix, until the GCC repo itself can be updated.

BUG=chromium:808822
TEST=emerge of cross-arm-none-eabi/gcc without this change fails;
with this change it passes.

Change-Id: I555a8a6ce315d49ece851a364ba23db09ccb0789
Reviewed-on: https://chromium-review.googlesource.com/900448
Commit-Ready: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>