unibuild: Don't require unit addresses
We currently require that the master config uses unit addresses when there
are multiple nodes with the same name. This means we need to specify a
'reg' property as well as #address-cells and #size-cells properties.

This can be confusing for people who are not familiar with device tree.
Add a dtc flag so that this is no-longer required.

BUG=chromium:796983
TEST=FEATURES=test sudo -E emerge chromeos-config-host

Change-Id: I4f2303cebccfb4ed71c18094f7da32b558c74ced
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/911977
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: C Shapiro <shapiroc@google.com>