chromeos-config: Disable yaml validation for a while
The YAML format of the master configuration is not actually used at
present. While we are working to update the schema it is annoying to have
to keep reef in sync. Adjust the ebuild to warn on validation problems,
rather than halting the build. This enables us to submit changes which
update the schema more easily.

We'll re-enable this once migration is complete.

BUG=chromium:796983
TEST=emerge-reef --nodeps chromeos-config
with this CL:
https://chromium-review.googlesource.com/c/chromiumos/platform2/+/900451/2
See that it builds now

Change-Id: Ic9db91779ca3278eb936759447b2e3519c87d919
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/902953
Reviewed-by: C Shapiro <shapiroc@google.com>