arc-build: Use -stdlib=libc++ instead of -lc++ for Clang
arc-cros-gralloc calls $CXX with flags that fail when extra libraries
are passed in.  For clang, the normal way to request linking to libc++
is to pass -stdlib=libc++ in $CXXFLAGS, so do that for clang builds.
GCC does not support -stdlib, so keep the old way for it.

This also makes newer versions of Clang (>= 5.0, arc-toolchain-master)
properly include its own C++ headers by default, so do not include GCC
C++ headers anymore in such case.

BUG=b:65189615
TEST=for i in {eve,kevin}{,-arcnext}; do emerge-$i arc-libdrm arc-cros-gralloc; done
TEST=for i in eve eve-arcnext; do emerge-$i arc-mesa; done
TEST=emerge-kahlee arc-llvm arc-mesa

Change-Id: Ic337a03a17a6925df68edbd43f843ff5c043c671
Reviewed-on: https://chromium-review.googlesource.com/923831
Commit-Ready: Tomasz Figa <tfiga@chromium.org>
Tested-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>