arc-build: Remove -lc++ from CXXFLAGS
arc-cros-gralloc calls $CXX with flags that fail when extra libraries
are passed in.  For clang, the normal way to request linking to libc++
is to pass -stdlib=libc++ in $CXXFLAGS, so we do that for clang builds.
gcc doesn't support -stdlib, but we can pass -lc++ in LIBS instead to
make sure it gets linked when needed.  This will fix arc-cros-gralloc
without breaking other packages.

BUG=b:65189615
TEST=rebuilt arc-llvm, arc-mesa, arc-cros-gralloc, and
     arc-amdgpu-addrlib

Change-Id: Ifbcf4714b413ecd6f8eeb91c4487c4f4839b1f26
Reviewed-on: https://chromium-review.googlesource.com/905241
Commit-Ready: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>