libchrome: inline FundamentalValue into base::Value
This moves the Value types which represent "primitives" and don't
have variable storage requirements into the base Value class and
moves a bunch of the logic for checking types into the base class
as well.

In addition, this uses subslots to ensure that the SDK libbrillo
is properly recompiled. Because we're removing a symbol, this
requires re-emerging host packages so they can pick up the alias
instead. However, subslots cause the SLOT to be added to ${OUT}
when they shouldn't be, so those should be stripped off while
constructing ${OUT}. Also, since protoc treats the colon as a
special character in some of its flags such as --cpp_out, we
replace it with double underscore so using non-zero slots won't
cause it to fail with an obscure error about unknown generator
options.

BUG=b:37434548
TEST=emerge, libbrillo unit tests

Change-Id: Iddd0239e1f46342671aa978bc3ebb87355f731ef
Reviewed-on: https://chromium-review.googlesource.com/924657
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>