arc-build: Replace CCLD override with something more self-contained
As per libtool manual [1], flags that libtool considers dangerous for
link process (in practice anything that is not on a hardcoded
whitelist), will be dropped from CCLD invocation. This is the reason why
our arc-build flags were not included originally. The same manual states
that

  "as a work-around you can prepend flags with -Wc, or -Xcompiler to
   allow them to be passed through to the compiler driver"

or

  "Another possibility is to add flags already to the compiler command
   at configure run time:

  ./configure CC='gcc -m64'"

Since multilib in portage already goes with the latter (-m32 is added
directly to CC) and I could get neither -Wc, nor -Xcompiler to work with
the flags we need, add our flags to CC and CXX as well. At the same
time, remove the custom src_compile() function from arc-mesa, added only
for the purpose of padding CCLD to emake.

BUG=b:65189615
TEST=emerge-kahlee arc-llvm
TEST=for i in eve eve-arcnext kahlee; do emerge-$i arc-mesa; done

Change-Id: I51cc5c19a8df41cf7bdd74d25a10533a7a3bc492
Reviewed-on: https://chromium-review.googlesource.com/923830
Commit-Ready: Tomasz Figa <tfiga@chromium.org>
Tested-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>