sys-devel/arc-toolchain-n: Stop including libLLVM in the toolchain
We do not use the libLLVM coming from Android anymore and so there is no
reason to keep including it in arc-toolchain-n. Remove anything related
to libLLVM from gather.sh script, so that it is no longer included in
toolchain packages.

As a consequence, there is no need to manually build anything from local
Android sources to deploy a new toolchain.

BUG=b:73424135
TEST=Run tryjobs for kevin-release, eve-release, soraka-release,
 kahlee-release and elm-release with a toolchain generated using
 the new script.

Change-Id: I176eecf553b52d2c17e0335c4afecd6b0964e041
Reviewed-on: https://chromium-review.googlesource.com/920046
Commit-Ready: Tomasz Figa <tfiga@chromium.org>
Tested-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Luis Hector Chavez <lhchavez@chromium.org>