sys-devel/arc-toolchain-*: Ensure correct permissions inside the tarball
The files inside the tarball must be readable (and executables
executable) for all users for the chroot to be able to use it. While
typically they come like this from their source locations, some
environments were found to reset the others permissions to o-rwx.

Make sure the behavior is consistent regardless of host configuration by
fixing up the permissions inside the gather.sh script. While at it,
reset the owner of all files to root:root, which is the common
convention for binary tarballs.

BUG=chromium:811217
TEST=Generate new arc-toolchain-master and
 sudo emerge arc-toolchain-master
 emerge-kevin-arcnext arc-libdrm
 emerge-eve-arcnext arc-libdrm arc-mesa
TEST=Generate new arc-toolchain-n and
 sudo emerge arc-toolchain-n
 emerge-kevin arc-libdrm
 emerge-eve arc-libdrm arc-mesa

Change-Id: Ia891754d61702baa6fe3ff878fe591adab873a16
Reviewed-on: https://chromium-review.googlesource.com/924105
Commit-Ready: Tomasz Figa <tfiga@chromium.org>
Tested-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Luis Hector Chavez <lhchavez@chromium.org>