cr50-scrips: do not try running factory scripts in prod images
The cr50-update.conf script is not supposed to run in the factory
environment, but to check if it is in the factory environment it tries
to source a file, which is not present in prod images, which in turn
causes initialization script failure.

Let's not try checking if we are in the factory environment if the
file is not present in the file system.

BRANCH=M65
BUG=b:73538102

TEST=verified that cr50-update.conf does not terminate with error in
     these scenarios:

     - normal mode boot, /usr/share/cros/factory_utils.sh is not
       present

     - normal mode boot, /usr/share/cros/factory_utils.sh is there,
       but the system is not if factory mode

    - /usr/share/cros/factory_utils.sh is there and the system is in
      factory mode (the appropriate message is added to
      /var/log/messages and the init script terminates)

    - recovery mode boot

Change-Id: I5a0f8db6fdbc40aa0431d0445053ccbe776b1c5c
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/924466
Reviewed-by: Andrey Pronin <apronin@chromium.org>