cros-workon: Use git ls-files for excluding files from copying
The --filter=":- <file>" switch of rsync, used currently, does not fully
handle .gitignore syntax and would behave incorrectly at least in
following two example cases:

a) negative matches

  gen*_pack.h
  !genX_pack.h

anything matching the first pattern, except anything matching the second
one too, should be ignored, but rsync reads the exclamation mark as a
part of file name pattern.

b) files checked in, but present in .gitignore - even though it's
questionable whether this is a correct repository setup, the affected
files are there in the checkout and so should be also included in the
source tree used for compilation.

Use git ls-files to query the list of ignored files instead and use the
output with --exclude-from rsync switch. This ensures that file
exclusion works exactly the same way as in git.

BUG=chromium:812979
BUG=chromium:812982
TEST=Create following directory stucture in src/third_party/arc-mesa
   x/y/ignored
   x/y/not_ignored
   y/ignored
   y/not_ignored
   y/.gitignore
 and put following contents in y/.gitignore
   ignored
 emerge-soraka arc-mesa copies all files except y/ignored.

Change-Id: I459a066c41f8effb2f7d655f324ab7af645bcd2e
Reviewed-on: https://chromium-review.googlesource.com/925222
Commit-Ready: Tomasz Figa <tfiga@chromium.org>
Tested-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>