chromeos-chrome: remove --strip-unneeded flag.
We plan to use eu-strip to replace binutils strip when using LLD
because of a bug in binutils strip.

The flag --strip-unneeded is not supported by eu-strip. But the
flag only works on object files, and it does not have effect on
executables and DSOs, as explained here:
https://www.technovelty.org/linux/stripping-shared-libraries.html

BUG=chromium:808262
TEST=cbuildbot caroline-chrome-pfq kevin-chrome-pfq

Change-Id: I48d8a70ce9a4d7b23e43a498d9c05a08459c160c
Reviewed-on: https://chromium-review.googlesource.com/929803
Commit-Ready: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>