Avoid runtime USE=test dependencies to trunks.
The purpose of these USE=test dependencies to trunks is to link
libtrunks_test.a, so they need not to be in RDEPEND.

I have not figured out the mechanism of crbug.com/813791 yet, but
those runtime dependency seem to refuse trunks to be re-emerged
without USE=test.

Also updates EAPI to 5.

BUG=chromium:813791
TEST=precq

Change-Id: Ie5a5875593370104f31560092603481fcc430383
Reviewed-on: https://chromium-review.googlesource.com/934107
Commit-Ready: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>