llvm: Add a fix to MemorySSA
Android was seeing issues where LLVM would incorrectly perform DSE in
some cases.

This fix should be sufficient to fix said DSE issues. Upstream is
tending a bit against this patch, because it seems we want to take
MemorySSA in a different direction (where it doesn't reason about
lifetimes at all). So, the "correct" fix will probably be a patch to
EarlyCSE. It's unclear how long it will take for consensus to be reached
+ a fix to EarlyCSE committed. Given that we branch in a week, and this
patch should be Good Enough(TM), Luis and I think it would be best to
carry it temporarily.

This patch is intended to make LLVM optimize code strictly less
aggressively, so it should be safe.

BUG=b:73127367
TEST=emerged llvm locally. cbuildbot chromiumos-sdk, peppy, terra,
peach_pit, kevin.

Change-Id: I9268b20a43ddff7a498921cc66f7b2b5b8552fd4
Reviewed-on: https://chromium-review.googlesource.com/932541
Commit-Ready: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>