cr50-scripts: script to support open box RMA verfication
This script will be invoked through crosh on a master Chrome OS device
in order to verify integrity of the AP and EC RO firmware on the
device under test.

The script will first verify if the DUT is running recent enough Cr50
firmware to perform the check, if not - the DUT Cr50 firmware will be
updated.

The RO verification database file is used to provide reference values
of AP and EC RO firmware hashes.

The script reports success or failure as text and exit value.

Install it along with other Cr50 shell scripts.

BRANCH=none
BUG=b:73668125
TEST=ran the script on the workstation to verify firmware on a Chrome
     OS device

Change-Id: I5a6ac80ee7db637b941ba96b5fcfbf703562984e
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/937658
Reviewed-by: Andrey Pronin <apronin@chromium.org>