Update Harfbuzz to 1.7.5 and up to 0bff6991
This is to match Harfbuzz on other platforms.

harfbuzz-1.7.5.324.tar.bz2 was obtained as following
from the harfbuzz git tree :

1. git checkout 0bff6991
2. sh autogen.sh
3. ./configure
4. make distcheck
5. rename the top directory to harfbuzz-1.7.5.324
6. make a tar ball.

Making a giant patch against harfbuzz-1.7.5.tar.bz to go up to
0bff6991 (324 CLs) does not work because applying the patch
changes Makefile.am (multiple of them) and configure.ac.
With the time stamp of those files updated by the patch application,
`make' looks for the same version of aclocal and  autoconf as used
in 'make distcheck' step above eventually leading to a build failure.

The simplest way is to upload the tar ball made above to the
CrOS common storage and use it directly without a patch.

BUG=818862
TEST=emerge-${BOARD} harfbuzz
TEST=cros tryjob -g 940450 chromiumos-sdk amd64-generic-full daisy-full

Change-Id: I8ba14b4b8349dfbcd3a487e21b7af96de59d1b6a
Reviewed-on: https://chromium-review.googlesource.com/940450
Commit-Ready: Jungshik Shin <jshin@chromium.org>
Tested-by: Jungshik Shin <jshin@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>