chromeos: turn on USE=forced-sandbox for mawk & sed
We currently always turn this on for mawk via a bashrc, so this doesn't
change behavior there.  It just makes it easier to sync with Gentoo and
drop our local patches (in the future).

For sed, this is new behavior.  We're disabling support for the e/r/w
commands because (1) we don't care about them and have no known legit
users and (2) they needlessly expose random shell scripts to attacks.

BUG=chromium:817993, chromium:818527
TEST=precq passes

Change-Id: I2173bb02fb271c6c31400efc792ad0b5b01e87b4
Reviewed-on: https://chromium-review.googlesource.com/949524
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: David Schneider <dnschneid@chromium.org>