chromeos-bootimage: Fix depthcharge target use
The current logic for building images for chromeos-bootimage iterates
over coreboot build targets. This is problematic when a different target
is being used for depthcharge; the ebuild script is unable to find the
correct file to pick up.

This changes the logic to iterate per firmware combination to ensure
that the right image is built. The files are current named after the
coreboot build target, but that will change once a "name" field is added
for build-targets in the master config.

BUG=chromium:819363
TEST=Tested a dtsi with depthcharge != coreboot target
CQ-DEPEND=CL:953124

Change-Id: I735d3e4833574943c6aeb812620e24bdbea920ba
Reviewed-on: https://chromium-review.googlesource.com/952205
Commit-Ready: Justin TerAvest <teravest@chromium.org>
Tested-by: Justin TerAvest <teravest@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Justin TerAvest <teravest@chromium.org>