libchrome: inline StringValue into base::Value
Two patches for ease of reviewing: one that removes JSONStringValue,
and one that inlines StringValue into base::Value. At this point all
of the Value APIs that would be broken by the uprev should be aligned
and so users of base::Value shouldn't have to make any more uprev
changes after the StringValue symbol is removed from users of
libchrome.

Also bump the subslot on libbrillo since this is going to break the
SDK again otherwise.

BUG=b:37434548
TEST=emerge, libbrillo unit tests, unit tests for ap packages on
  whirlwind

Change-Id: I043e39318e397cee54416b73978dc14f039b57dd
Reviewed-on: https://chromium-review.googlesource.com/953329
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>