libchrome: Filter signal by the sender we are interested in only.
This adds sender='<name>' filter as the match rule for subscribing to
signal in ObjectProxy. This is needed to reduce the unnecessary polluted
log of rejecting some signals. The manual sender verification is still
needed because as mentioned in
https://bugs.chromium.org/p/chromium/issues/detail?id=507206#c14
service_name_ might be empty and sender='' means to not filter by sender
at all.

This is a cherry-pick of http://crosreview.com/958324.

BUG=chromium:812468
TEST=Tested with newblue daemon

Change-Id: Ia9111559799a1354006b3599dca43a6b0c30bdd8
Reviewed-on: https://chromium-review.googlesource.com/958039
Commit-Ready: Sonny Sasaka <sonnysasaka@chromium.org>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Satoru Takabayashi <satorux@chromium.org>