net-print/ippusbxd: Patching ippusbxd to hold usb interfaces
BUG=chromium:768606
TEST=Manually tested on chromebook

The current version of ippusbxd claims a USB interface and subsequently
releases it for each connection that CUPS makes to the printer.
Unfortunately, some printers misbehave when one interface is released
while a different interface is in the middle of transferring data.

This change will have ippusbxd claim all of the needed USB interfaces at
the time of startup, and only release them once the program has
completed.

Change-Id: Iae44d3a8ff5112656578ce99c75f467c69624945
Reviewed-on: https://chromium-review.googlesource.com/965242
Commit-Ready: David Valleau <valleau@chromium.org>
Tested-by: David Valleau <valleau@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>