net-print/ippusbxd: Changing communication code to be bidirectional
BUG=chromium:768606
TEST=Tested manually on device using a handful of different printers
CQ-DEPEND=CL:965242

Communication between the printer and CUPS was implemented in a way
which only read from one end at a time. This change uses two threads
which are responsible for reading from one end and sending it to the
other, which removes the need to make any assumptions about whose "turn"
it is to transfer data.

Change-Id: Ia02167d67930d276c86e828614e80bcd1c483a7f
Reviewed-on: https://chromium-review.googlesource.com/969789
Commit-Ready: Sean Kau <skau@chromium.org>
Tested-by: David Valleau <valleau@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>