python: disable build overrides for the sdk itself
We deployed this var override so we can cross-compile python packages,
but it's not playing well with multitarget python code (since this func
runs once, but the python probes multiple versions).  Drop it for the
sdk since the reason for the workaround isn't necessary, and we want to
enable multiple versions of python there.

We'll need to revisit this if we want to do multiple versions of python
in the board, but we also need to do custom patches on python 3.x first.

BUG=chromium:736322
TEST=`emerge dbus-python` works correctly for the sdk

Change-Id: Ie24782028f2a00700cc4bb5c28f2010d0d14ad7d
Reviewed-on: https://chromium-review.googlesource.com/969903
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>