infineon-firmware-updater: Add owner auth update type.
This adds a new update type to infineon-firmware-updater that use the
well-known secret (all zeros) as the owner password. The type is
selected via -update tpm12-ownerauth.

Use this new mode in tpm-firmware-updater when running in normal mode,
the TPM is already owned, and an auth check passes.

BUG=chromium:788719
TEST=unit tests

Change-Id: Ic5554d5caa528943a5175260e57c1d625f583260
Reviewed-on: https://chromium-review.googlesource.com/978172
Commit-Ready: Mattias Nissler <mnissler@chromium.org>
Tested-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>