Add edid-decode package to chromeos-base/chromeos-chrome ebuild files
Added edid-decode package to chromeos-base/chromeos-chrome ebuild files;
this will allow in a subsequent CL to run it as part of chrome:system
to include the parsed EDID blobs in the system report (not that the
blob is already available, albeit not parsed, under "modetest").
The binary size increase is minimal (~8KB I think).

Change-Id: I1e5b62c551a4d330d06c004c13b6a1f97f6d2630
Bug: 821393
Reviewed-on: https://chromium-review.googlesource.com/974653
Commit-Ready: Miguel Casas <mcasas@chromium.org>
Tested-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>