llvm: Update wrappers for multiple llvm installations.
Currently only one version of llvm is installed in the chroot.
Update the host and target wrappers to support using a llvm
installation in a custom location.
A new wrapper option "-Xclang-path=" is added to specify the location
of the custom clang installation.

Any package that needs to use a different clang than the default
should specify this option.

BUG=chromium:773887
TEST=Chromiumos-sdk, kevin, peppy tryjobs passed.

Change-Id: If865d3ce843a0490c855d29484f469ce4fcbb94b
Reviewed-on: https://chromium-review.googlesource.com/982397
Commit-Ready: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>