[clang-tidy] Update log parsing script.
This adds flags allowing the user the specify the input and output
files/directories; tries to add the board name to the output directories,
and the package name and date to the output file names (making it
possible to run clang-tidy on multiple packages, not just Chrome).

BUG=chromium:826591
TEST=Ran script with various new input/output designations and
they all seemed to work.

Change-Id: I773ba76ed5928adc891968e337348f5cefc482d1
Reviewed-on: https://chromium-review.googlesource.com/985313
Commit-Ready: Caroline Tice <cmtice@chromium.org>
Tested-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Caroline Tice <cmtice@chromium.org>