chromeos-cr50-scripts: when booting from USB do not restart for Cr50 updates
System restart when booting from a removable device is quite annoying,
and it is bound to happen if the Cr50 image on the removable device is
newer than the image currently running on H1.

This patch prevents restarts when system boots from a removable device.

This patch also creates a situation where the currently running Chrome
OS version potentially is expecting a much newer Cr50 than the one
currently running. But this is an inconvenience we are prepared to
live with.

Also adding a timeout after issuing the 'reboot' command to ensure
that this script's processing stops until reboot.

BUG=chromium:709611
TEST=verified that when booting from a USB flash drive or SD card, the
     Cr50 gets updated, but the device is not rebooted for this reason.
     The new Cr50 image takes over after the user reboots the device.
     verified that cr50-result still reboots the device during startup
     if the update failed during chromeos-install

Change-Id: Ic349f52637bafeb89cb971e36a56d53b8abd4e70
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/985613
Commit-Ready: Mary Ruthven <mruthven@chromium.org>
Tested-by: Mary Ruthven <mruthven@chromium.org>