chromeos-cr50: uprev versions: prod to 0.3.4, pre-pvt to 0.4.4
This patch is part of a set of patches which uprevs all Chrome OS
devices to prod version 0.3.4 and pre-pvt version 0.4.4.

The 0.4.4 version is the previously released 0.2.2 re-signed with a
new manifest to make sure the released pre-pvt version numbers stay
above the released version numbers.

The new image brings in a lot of bug fixes and new features, too many
to mention them all, most notable are:

  - fixes for boot failures caused by problems synchronizing reset
    state with the AP

  - RMA unlock allowing to enable device reimaging in controlled
    environment

  - open box RMA verification allowing to use the Cr50 to verify that
    it's master AP RP firmware has not been modified after
    manufacturing.

The new prod image version string is

  Build:   0.3.4/cr50_v1.9308_87_mp.30-a9768e0
           tpm2:v0.0.291-af944a7
           cryptoc:v0.0.8-6283eee
           2018-03-23 18:20:09 @build22-m2.golo.chromium.org

The new pre-pvt version string is

  Build:   0.4.4/cr50_v1.9308_B.334-78f874b
           tpm2:v0.0.291-af944a7
           cryptoc:v0.0.8-6283eee
           2018-02-20 23:24:38 @build21-m2.golo.chromium.org

CQ-DEPEND=CL:*589987
CQ-DEPEND=CL:*598563
CQ-DEPEND=CL:*598097
BUG=b:72295419
TEST=verified that with dependent patches applied boards bob, coral,
     eve, fizz, reef, scarlet, and soraka install the same pair of
     Cr50 images, 0.4.4 for pre-pvt and 0.3.4 for prod

Change-Id: I8db8f7be84fcf859efdbc6001130a29774348685
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/987398
Reviewed-by: Nick Sanders <nsanders@chromium.org>