chromeos-base: Remove old 99-usb-printer.rules.
Remove an ancient udev rule that looks like it was used to
notify Chrome when a USB printer was connected (so Chrome
could display an error?).

Chrome was notified by emitting a signal on the
org.chromium.LibCrosService D-Bus service, which will soon
be removed. I can't find anything that consumes this signal
anymore.

BUG=chromium:692246,chromium:218515
TEST=none

Change-Id: If736ee97f9dc658e2458f9016eb6dabbda2320e6
Reviewed-on: https://chromium-review.googlesource.com/1001094
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>