wpa_supplicant: give DBUS perms to 'shill' user
This CL is in preparation to run the shill process tree as a non-root
user.

Shill needs DBUS access to wpa_supplicant. Currently it runs as root,
but soon it will run as non-root user 'shill'. All code that will run as
shill currently runs as root, so this isn't a security regression.

BUG=chromium:649417
TEST=tested as part of larger shill sandboxing debug CL

Change-Id: If05e85ed57c0d57686d4677ec450e3b3d0056f26
Reviewed-on: https://chromium-review.googlesource.com/1007491
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>